requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
distro==1.9.0
filelock==3.16.0
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-anthropic==0.1.23
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langsmith==0.1.117
MarkupSafe==2.1.5
multidict==6.1.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
setuptools==72.1.0
sniffio==1.3.1
SQLAlchemy==2.0.34
tenacity==8.5.0
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wheel==0.43.0
yarl==1.11.1
